SNOW-2972472 moved tests suite for caching into mod.rs
